[Remove] viewModel에서 input, state, error 구현했던 추가 파일 제거하고 원본 파일로 통합
